Create new app Authors, move Author Class to new app.
Split Authors and Books funcionality to two diffrent apps.
