+++ MainApp/main_app/authors/admin.py
@@ -1,23 +1,15 @@
 from django.contrib import admin
 
-from .models import Book, Author
-
+from google_book_api.models import Book
+from .models import Author
 
 # Register your models here.
-
 class AuthorToBookInline(admin.StackedInline):
     model = Book.authors.through
-
-
-class AdminBook(admin.ModelAdmin):
-    filter_horizontal = ('authors',)
-    prepopulated_fields = {"slug":["api_book_id"]}
-
 
 class AdminAuthor(admin.ModelAdmin):
     inlines = [
         AuthorToBookInline
     ]
 
-
-admin.site.register(Book, AdminBook)
+admin.site.register(Author, AdminAuthor)
