+++ MainApp/main_app/google_book_api/models.py
@@ -1,19 +1,9 @@
 from django.db import models
 from django.utils.text import slugify
 
+from authors.models import Author
 
 # Create your models here.
-
-
-class Author(models.Model):
-    name = models.CharField(max_length=100)
-    birth_date = models.DateField()
-    death_date = models.DateField(blank=True, null=True)
-    biography = models.TextField(blank=True, null=True)
-
-    def __str__(self):
-        return self.name
-
 
 class Book(models.Model):
     slug = models.SlugField(unique=True)
