+++ MainApp/main_app/main_app/settings.py
@@ -35,7 +35,8 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'google_book_api'
+    'google_book_api',
+    'authors'
 ]
 
 MIDDLEWARE = [
